Move detectConsentManager from Tour to SitesManager API
